redux pt 1
