+++ redux/store.js
@@ -1,0 +1,34 @@
+import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit'
+import { quotesReducer } from '../features/quotesSlice';
+import {
+    persistStore,
+    persistCombineReducers,
+    FLUSH,
+    REHYDRATE,
+    PAUSE,
+    PERSIST,
+    PURGE,
+    REGISTER
+} from 'redux-persist';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+
+const config = {
+    key: 'root',
+    storage: AsyncStorage,
+    debug: true
+}
+
+export const store = configureStore({
+    reducer: persistCombineReducers(config, {
+        quotes: quotesReducer
+    }),
+    middleware: (getDefaultMiddleware) =>
+        getDefaultMiddleware({
+            serializableCheck: {
+                ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER]
+            }
+        })
+});
+
+export const persistor = persistStore(store);
+
