+++ shared/quotes.js
@@ -1,0 +1,7 @@
+export const QUOTES = [
+    {
+        id: 0,
+        text: 'Adversity causes some men to break, others to break records, am i rite?',
+        author: 'William Arthur Ward'
+    }
+];
